+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Listing assets with no results returns an empty array.
 - Fixed a panic that could occur when creating resources in a namespace that
 does not exist.
+- Add a timeout to etcd requests when retrieving the nodes health
 
 ## [5.15.0] - 2019-11-18